Troop 233 Tree Recycling script
This is the splitter script (and the ipynb it was developed from) which
will take the master .xlsx (from google sheets) and split it into
separate files for each route.

There are some assumptions about the format of the google sheets
embedded in this script, so be careful when renaming columns.