create another container for 'printing'
It looks like our wkhtmltopdf doesn't understand page breaks along the
cell lines.  Yet when we open the html in chrome and try the same it
does it perfectly.

So we're going to explore using another container with chromium headless
that we can print from.  This establishes the 'printer' container with
Node, Chromium, and Puppeteer from
https://blog.risingstack.com/pdf-from-html-node-js-puppeteer/