New Procedure for pdf generation
This changes the pdf generation from wkhtml2pdf which was not breaking
the pages correctly.

The new process is much slower as it uses a chromium browser to load the
html pages and render the pdf.  This uses the chromium browser to break
the page, which it does admirably.

This uses a second docker image for chromium and the node script to
print with it.  Make sure to ./dockerbuild again.